REDIS added
Microsoft.Extensions.Caching.StackExchangeRedis

Microsoft.Extensions.Caching.Memory

Create a service implementing the ICacheService interface

This will be on the infrastructure layer.

Since this is not a test case, no repository classes are needed, as we are not accessing any entities.

Inject the Redis dependency in the infrastructure extension.

AppSettings configuration

If the Redis connection string is disabled, use in-memory caching (local RAM).

If it is enabled, connect to the remote Redis server.